Simplify headless context handling
