+++ modules/webgl/src/init.js
@@ -62,6 +62,16 @@ if (!global.luma) {
   };
 }
 
+// Ensures that WebGL2RenderingContext is defined in non-WebGL2 environments
+// so that apps can test their gl contexts with instanceof
+// E.g. if (gl instanceof WebGL2RenderingContext) { }
+class WebGL2RenderingContextNotSupported {}
+global.WebGL2RenderingContext = global.WebGL2RenderingContext || WebGL2RenderingContextNotSupported;
+
+// Ensure that Image is defined under Node.js
+class ImageNotSupported {}
+global.Image = global.Image || ImageNotSupported;
+
 export {global};
 export {lumaStats};
 export default global.luma;
